[Kernel] [Pagination] Check Log Segment Changes between Page Requests (#4926)
<!--
Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines:
https://github.com/delta-io/delta/blob/master/CONTRIBUTING.md
2. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP]
Your PR title ...'.
  3. Be sure to keep the PR description updated to reflect all changes.
  4. Please write your PR title to summarize what this PR proposes.
5. If possible, provide a concise example to reproduce the issue for a
faster review.
6. If applicable, include the corresponding issue number in the PR title
and link it in the body.
-->

#### Which Delta project/connector is this regarding?
<!--
Please add the component selected below to the beginning of the pull
request title
For example: [Spark] Title of my pull request
-->

- [ ] Spark
- [ ] Standalone
- [ ] Flink
- [x] Kernel
- [ ] Other (fill in here)

## Description

<!--
- Describe what this PR changes.
- Describe why we need the change.
 
If this PR resolves an issue be sure to include "Resolves #XXX" to
correctly link and close the issue upon merge.
-->
Check if the log segment has changed between page reqeusts. 

## How was this patch tested?

<!--
If tests were added, say they were added here. Please make sure to test
the changes thoroughly including negative and positive cases if
possible.
If the changes were tested in any way other than unit tests, please
clarify how you tested step by step (ideally copy and paste-able, so
that other reviewers can test and check, and descendants can verify in
the future).
If the changes were not tested, please explain why.
-->
PaginationContextSuite.scala.

## Does this PR introduce _any_ user-facing changes?

<!--
If yes, please clarify the previous behavior and the change this PR
proposes - provide the console output, description and/or an example to
show the behavior difference if possible.
If possible, please also clarify if this is a user-facing change
compared to the released Delta Lake versions or within the unreleased
branches such as master.
If no, write 'No'.
-->

+++ kernel/kernel-api/src/test/scala/io/delta/kernel/internal/PageTokenSuite.scala
@@ -39,7 +39,7 @@ class PageTokenSuite extends AnyFunSuite with MockFileSystemClientUtils {
   private val TEST_TABLE_PATH = "/path/to/table"
   private val TEST_TABLE_VERSION = 5L
   private val TEST_PREDICATE_HASH = 123
-  private val TEST_LOG_SEGMENT_HASH = 456L
+  private val TEST_LOG_SEGMENT_HASH = 456
 
   private val expectedPageToken = new PageToken(
     TEST_FILE_NAME,
@@ -79,7 +79,7 @@ class PageTokenSuite extends AnyFunSuite with MockFileSystemClientUtils {
     assert(row.getString(4) == TEST_TABLE_PATH)
     assert(row.getLong(5) == TEST_TABLE_VERSION)
     assert(row.getInt(6) == TEST_PREDICATE_HASH)
-    assert(row.getLong(7) == TEST_LOG_SEGMENT_HASH)
+    assert(row.getInt(7) == TEST_LOG_SEGMENT_HASH)
   }
 
   test("E2E: PageToken round-trip: toRow -> fromRow") {
